requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Google Calendar Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# PayPal Integration
paypalrestsdk==1.13.3

# Environment Management
python-dotenv==1.0.0

# Development Tools
black==23.9.1
flake8==6.1.0
